api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["manhtd98 <manhtd6932@seta-international.vn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.19"
celery = "^5.3.1"
fastapi = "^0.100.1"
uvicorn = "^0.23.2"
redis = "^4.6.0"
nanoid = "^2.0.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
opentelemetry-api = "^1.19.0"
opentelemetry-sdk = "^1.19.0"
opentelemetry-instrumentation-fastapi = "^0.40b0"
opentelemetry-exporter-zipkin = "^1.19.0"
opentelemetry-exporter-otlp = "^1.19.0"
loguru = "^0.7.0"
protobuf = "3.20.1"
ecs-logging = "^2.0.2"
fluent-logger = "^0.10.0"
msgpack = "^1.0.5"
python3-logstash = "^0.4.80"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
